+++ src/Interfaces/HasValidate.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ReallifeKip\ImmutableBase\Interfaces;
+
+interface HasValidate
+{
+    public function validate(): bool;
+}